[KYLIN-4780] Fixed some unreasonable design
